Crate libsodium_seeded_prng

Implementation of libsodium's randombytes which allows a seed to be passed.

Structs

Seed
A simple wrapper used to seed the RNG which prints its value on destruction if the current thread is panicking.

Enums

Error
Error types

Functions

get_rng
Return a copy of the thread-local RNG pointer

get_seed
Returns a copy of the current RNG seed.

implementation_name
Returns the name of this libsodium randombytes implementation.

init
Sets libsodium randombytes to this implementation and initialises libsodium.

random_bytes
Returns a vector of random bytes of length size.

random_u32
Returns a random u32.

random_u32_uniform
Returns a random u32 between 0 and upper_bound (excluded).